requirements.txt
black==23.7.0
pytest==7.4.0
pytest-bdd==6.1.1
pytest-dotenv==0.5.2
pytest-xdist==3.3.1
requests==2.31.0
allure-pytest==2.13.2
pytest-html==3.2.0
reformat-gherkin==3.0.1
